% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CA26.R
\name{CA26}
\alias{CA26}
\title{Loading, cleaning, and calculating performance ratios for data analysis.
Returns a clean data set with relevant metrics for analysis.}
\usage{
CA26()
}
\description{
Merges all datasets from ws.data.
}
\details{
Calculates
}

