% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{table1}
\alias{table1}
\title{Computing future performance metrics for table1}
\usage{
table1()
}
\description{
Calculating and grouping stocks based on their historical relative strengths
and comparing 4 and 26 week future returns.
}
\details{
Resulting table is saved in the data file.
}

