% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2.R
\name{table2}
\alias{table2}
\title{Computing future performance metrics for table2}
\usage{
table2()
}
\description{
Calculating and grouping stocks based on their historical relative strengths, sub-grouping
by variation ranks, and comparing 4 and 26 week future returns.
}
\details{
Resulting table is saved in the data file.
}

