% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3.R
\name{table3}
\alias{table3}
\title{Computing future performance metrics for table3}
\usage{
table3()
}
\description{
Calculating and grouping stocks based on their historical relative strengths, sub-grouping
by weekly market performance, and comparing 4 and 26 week future returns.
}
\details{
Resulting table is saved in the data file.
}

