% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table4.R
\name{table4}
\alias{table4}
\title{Computing future performance metrics for table4}
\usage{
table4()
}
\description{
Calculating and grouping stocks based on their historical relative strengths
and comparing 52 week returns.
}
\details{
Resulting table is saved in the data file.
}

