% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table5.R
\name{table5}
\alias{table5}
\title{Computing future performance metrics for table6}
\usage{
table5()
}
\description{
Calculating and grouping stocks based on their historical relative strengths, sub-grouping
by variation ranks, and comparing 52 week future returns.
}
\details{
Resulting table is saved in the data file.
}

