% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table6.R
\name{table6}
\alias{table6}
\title{Computing future performance metrics for table2}
\usage{
table6()
}
\description{
Calculating and grouping stocks based on their historical relative strengths, sub-grouping
by market performance, and comparing 52 week future returns.
}
\details{
Resulting table is saved in the data file.
}

